Rewrite MDX links with query strings
Previously, given a link like [this](../mypage.mdx?scope=cloud),
remark-links would fail to rewrite the link href. This was because
the regexp used to check MDX links would only check for a # character
or the end of the line after the .mdx? file extension. This edits the
regexp to change the behavior and eventually enable us to link to other
scopes.