Add ScopeSwitchTile
While working on adjusting the visibility of docs nav bar entries based
on scope (https://github.com/gravitational/teleport/issues/11383), I
realized that some nav bar sections are filled with irrelevant pages,
but a user might want to read the introduction of that section.

For example, the Enterprise section is irrelevant to OSS users, but
these users may still be interested in learning more about Teleport
Enterprise.

One possibility would be to leave these section as-is for readers with
an unintend scope. However, these readers would see edition warnings in
the pages within a section unless they changed the value of the scope
picker.

As an alternative, we can hide scope-irrelevant content from the nav
bar menu and section introduction pages, but include those pages, and
add a prominent link that enables the reader to change scope.

I've suggested the ScopeSwitchTile as a way of doing this. When a user
clicks the ScopeSwitchTile, the current scope changes to a caller-
configured value.

We would use a ScopedBlock to hide menu items in section introduction
pages from users with scopes that are not related to the section. We
would then use a ScopeSwitchTile to enable the user to switch scopes
and see the related content.