Add a ScopeLink component
In the docs, there are various guides that are entirely irrelevant for
a user of a particular scope. In these cases, we usually include a
warning at the top of each guide that indicates that the guide is only
relevant for Cloud, Open Source, or Enterprise users. Rather than force
the user to guess that they need to change the value of the scope
picker, the ScopeLink component enables us to add an inline link that
changes the scope of the page. For example:

```html
<Notice
  type="warning"
  scope={["oss"]}
>

  This guide requires Teleport Cloud or Teleport Enterprise.

  View this guide as a <ScopeLink scope="cloud">Teleport Cloud
  user</ScopeLink> or <ScopeLink scope="enterprise">Teleport
  Enterprise user</ScopeLink>.

</Notice>
```