Throw a linter error on "bash" code blocks.
We introduced code blocks with language "code" in order to apply
custom functionality via remark. For older guides that don't include
the "code" language in code blocks, as well as docs contributors who
may not know about the "code" language, it would be helpful to throw
errors on "bash" blocks in the linter.