Adds CI/CD scripts for docs image
Adds a build script that will construct and publish the `docs` image to
`quay.io`. This build script will be triggered by Google Cloud Build
on a push to `main`.

The upshot of this script is that the image used by Teleoprt CI to
run its doctests will be automatically updated every time a PR is
merged in this repository.

See-Also: gravitational/teleport#13457