Check config.json URLs for corresponding files (#123)
* Check config.json URLs for corresponding files

Fixes #101

When loading the `docs/config.json` file for each version of the docs
site, we have some logic in `normalizeDocsUrl` that validates URLs
(i.e., checks if the URL has a trailing slash) and adds a
version-specific path segment to the URL path.

This changes `normalizeDocsUrl` to also check if a file exists at the
given URL path. This way, docs site builds fail if (a) a navigation
entry exists for a nonexistent docs page; or (b) a redirect points to a
nonexistent docs page.

Unfortunately, throughout the code that handles `config.json`, there is
logic that assumes the file lives in
`content/<version>/docs/config.json`. To test this change on a test data
directory, rather than an actual docs site directory, we would need to
refactor a lot of `server/config-docs.ts` and the code that imports from
it.

As a result, in order to test only `normalizeDocsUrl`, I had to make it
an exported function. I also had to make assumptions in the tests that
there would be a `content` directory populated with per-version
subdirectories.

* Respond to PR feedback

* Refactor the config.json checker

We can't guarantee that our CI server will have loaded this repo's git
submodules before running tests. This change refactors the
config.json checker so we can test it with any docs content directory,
e.g., the `server/fixtures` directory.

This also provides better separation of concerns between the
URL normalization code (which doesn't need to know about a directory
tree) and the file path checker code.