Make the config.json error message more explicit
We recently edited the docs engine to check the config.json for each
docs version to ensure that navigation entries and redirects do not
point to nonexistent pages. However, the error message was a bit
cryptic, causing some confusion about what was causing the error.

This change edits the error message to clarify that it has to do with
the config.json file for a specific version of the docs.