Allow more paths in partial link resolution
Fixes #167

`handlePartialLink` currently hardcodes `docs/pages` as the location of
all docs pages, and uses this file path segment to construct the absolute
path of the MDX file that includes a partial.

This change adds the root directory of all partials as a parameter of
`handlePartialLink` and uses that instead. Since this root directory is
configurable outside `remark-includes`, this change lets us add unit
tests for `handlePartialLink`.

This change also handles relative path resolution for image and link
definition URLs. The current version of `handlePartialLink` only handles
links paths. As a result, this renames `handlePartialLink` to
`handleURLPath`.