Remove `hideInScopes` and related logic.
Closes #87

It is currently possible to add the `hideInScopes` field to a navigation
entry in the docs config file. This field indicates in which values of
`DocsContext.scope` a link to the page will be visible in the navigation
sidebar. This caused confusion with users, so we stopped including the
`hideInScopes` property. This change removes the `hideInScopes`
functionality entirely.