Specify the remark config path in markdown-lint (#232)
Closes #231

The `markdown-lint` command defined in `package.json` does not define a
config file, meaning that `remark` will search for a config file
automatically.

After we merged the `gravitational/webapps` code with
`gravitational/teleport`, `gravitational/teleport` started to include
multiple `package.json` files. This interferes with `remark`'s config
search, causing `remark` to ignore our plugin config. As a result, the
linter passes when it should be returning errors.

By specifying a `remark` config file explicitly in `markdown-lint`, we
can ensure that `remark` uses the correct config regardless of the
content of the versioned `content` submodules. Since
`gravitational/docs` is decoupled from its submodules, we should not
expect a given version of a submodule to necessarily play nicely with
the tools we run in the parent module.