Lint incorrect messaging in the docs
Closes #154

One challenge for maintaining the docs is ensuring that we maintain
consistent terms for different Teleport products and components.

This change adds a linter to prevent incorrect usage of terms and
suggest alternatives with an explanation.

The configuration file, which must live in `docs/messaging-config.json`
within a `gravitational/teleport` repo, looks like this, an array where
each element is an object that specifies a rule:

```json
[
    {
    	"incorrect": "\\W(machine id|(database|app(lication)?|desktop|kubernetes|ssh|discovery) service)\\W",
    	"correct": "capitalize the names of Teleport services",
    	"explanation": "We want to ensure that product names are consistent across the docs"
    }
]
```

The `incorrect` field is a regular expression indicating a pattern to
avoid. The config above would show a warning in the linter on, for
example, `machine id`, `database service`, and `app service`. The
`correct` and `explanation` fields are used to print error messages.

The intention is for docs authors to be able to adjust the linter config
along with the docs content.