Add a spellcheck yarn job
Closes #115

Add a `spellcheck` yarn script that we can run in the
`gravitational/teleport` CI job for linting docs changes.

This change assumes that there is a `cspell.json` configuration file at
the root of each supported docs version that includes words to ignore
during the spellcheck.

Note that this command uses `find` with `xargs` instead of `find` with
the `-exec` flag because the former returns a nonzero exit code if one
of the `-exec` runs fails.