Lint incorrect messaging in the docs (#238)
* Lint incorrect messaging in the docs

Closes #154

One challenge for maintaining the docs is ensuring that we maintain
consistent terms for different Teleport products and components.

This change adds a linter to prevent incorrect usage of terms and
suggest alternatives with an explanation.

The configuration file, which must live in `docs/messaging-config.json`
within a `gravitational/teleport` repo, looks like this, an array where
each element is an object that specifies a rule:

```json
[
    {
    	"incorrect": "\\W(machine id|(database|app(lication)?|desktop|kubernetes|ssh|discovery) service)\\W",
    	"correct": "capitalize the names of Teleport services",
    	"explanation": "We want to ensure that product names are consistent across the docs"
    }
]
```

The `incorrect` field is a regular expression indicating a pattern to
avoid. The config above would show a warning in the linter on, for
example, `machine id`, `database service`, and `app service`. The
`correct` and `explanation` fields are used to print error messages.

The intention is for docs authors to be able to adjust the linter config
along with the docs content.

* Respond to PR feedback and define a config

- Load the config from gravitational/docs
- Define some initial messaging rules

* Make messaging linting aware of page parts

- Make it possible to configure messaging linting for certain parts of
  the docs.

  This way, we can ignore "Server Access" etc. in titles while
  disallowing it in comments, headers, body text, and the page
  description. "Server Access" etc. make sense in title case even when
  we are not using these terms as product names, since significant words
  are capitalized. But these terms don't work with the new messaging
  conventions where we expect regular nouns not to be capitalized.

- Mention the part of a docs page where an issue occurred in warning
  messages. This way, it's clearer to a user that they should fix a code
  comment, title, description, etc. For parts of a docs page that aren't
  regular text, the line numbering can often be deceptive, so this makes
  issues easier to fix.