Add note re: messaging linter line accuracy
Line numbers in messaging linter warnings are not accurate if the
violation takes place in a partial. The issue is a bit complex and seems
to have to do with the way `remarkIncludes` manipulates the MDAST.
Before we can address gravitational/docs#273, this change adds a note
that line numbers may not be correct for violations that take place
inside partials.