Add a transformToAST test
Do this instead of testing a custom series of processors in
`remark-includes.test.ts`. Doing this because transformToAST already
returns an AST, so in theory makes for a cleaner test. Also, I think the
eventual solution will be to recursively call transformToAST on each
include, so doing this lets us test this approach directly.

Note that this is currently broken since transformToAST hardcodes a
number of expectations about the structure of the content directory
structure.