Lint absolute links to docs pages (#304)
* Lint absolute links to docs pages

Fixes #156

This change edits `remark-links` to flag absolute links to pages within
the docs, e.g., `https://goteleport.com/docs` or `/docs`. To do so, it
adds a `lint` config option to `remark-links` that checks for URLs to
the paths within the docs site that are not relative paths to an MDX
file.

An alternative would be to create a separate linter for absolute docs
link paths, instead of branching within `remark-links`. The advantage of
the approach within this change is keeping the work of processing
Markdown links within a single remark plugin.

This also catches absolute docs site links within the `href` properties
of MDX components.

* Separate out the absolute docs link linter

This makes the code cleaner than including the linter within the same
plugin as `remarkLinks`, which modifies docs links.

* Remove dead code