Fix Var component copy/paste
Closes #336

When the docs engine renders a code snippet, it either renders a
`Snippet` (for `code` blocks) or a `Pre` (for other kinds of snippets).
The `Snippet` component wraps a `Pre`. Non-`code` snippets get processed
by `rehype-highlight`, which transforms the code snippet into a
`<pre><code></code></pre>` with `hljs` classes.

The `Pre` component calls a `toCopyContent` function to extract text
from its child components and write it to the clipboard, and the
`Command` component (which represents a command within a `code` snippet)
uses this function to copy individual command lines. However,
`toCopyContent` assumes that its child nodes have a particular
structure, which is easy for components that invoke `Pre` to violate.

Solution:

Pass an array of class names to `toCopyContent` so it doesn't need to
anticipate a specific HTML structure. `toCopyContent` fetches the text
content of each HTML node it finds that matches the target class names.
This works for both `code` and non-`code` snippets.