Get all Storybook tests to pass
The CopyButton story in `Command.stories.tsx` calls `setup()` on
`userEvent`, assuming this is the `userEvent` from React Testing
Library. However, this code imports `userEvent` from
`@storybook/testing-library`, which does not support the `setup()`
method.

To get the test to pass, this change uses the
`replaceClipboardWithCopyBuffer` function, which replaces
`window.navigator.clipboard` without the additional dependency.