Copy only commands in command snippets
Fixes #121

The `CodeLine` component represents a string within a `code` snippet
that is not a command. This is often an example of a command's output.

However, for code snippets consisting of, say, example YAML, `CodeLines`
represent the value we want a user to copy. This use case enables
example values to include `Var` components, and we'll likely deprecate
it once we enable support for `Var`s in syntax-highlighted code blocks.

Since a `CodeLine` can show up in both of these use cases, we need to
make sure that they are copied when there are no commands in a `code`
snippet and _not_ copied when there are commands. This is because, when
there are commands in a snippet, the intention is for a reader to copy
only the commands--the output would be what the reader sees when they
enter the commands into their terminal.

This change adds branching in the `Pre` component, which handles the
copy/paste logic for `Snippet`, to copy `CodeLine`s only when there are
no `CommandLine`s in the snippet.