Render TabItems with identical labels correctly (#349)
Fixes #141

The `TabItemList` component renders a `TabItem` if its label is
selected, but does not evaluate whether a `TabItem`'s `options` property
is selected. This means that, if two `TabItem`s have different options
but identical labels, both `TabItems` will render.

This change edits `TabItemList` to evaluate both a `TabItem`'s `options`
and its `label` to determine whether to render it.

In order to get the Storybook tests to work, this also edits the
`Dropdown` component to remove the "portal" option from the Reach UI
`ListBoxPopover`. This makes the popover available to Storybook.