Render dropdown popover above code snippets (#360)
Fixes #356

The earlier change #349 edited the `Dropdown` component to set `portal`
to `false`, preventing the component from creating a separate React
root. While this made it easier to test the Dropdown component within
Tabs, also caused the component to render below code snippets.

This change edits the CSS `z-index` of the Dropdown popover so it
displays above code snippets. It leaves the `portal` prop as `false`.