Support default values in the Var component (#355)
Fixes #190

Support default values in the `Var` component that are distinct from the
component's `name`. This way, we can have relatively simple `Var` names
but more involved default values.

While expanding the `Var` component, I also removed some optimization
callbacks that I'm not sure we really need yet in order to simplify the
`Var` code, and changes the `isGlobal` prop to a boolean to simplify
operations that use it (no docs pages use `isGlobal` yet, so it's safe
to make this change).

Also condenses `setField` and `addField` into a single `putField`
function, and declares a separate `setInitial` function for the initial
value of a given `Var`.