Render empty lines in example `code` output
Fixes #103

`remark-code-snippet` splits each line in the content of a `code`
snippet and produces either a `CommandLine` or a `CodeLine` for each
line. If the line is example output (i.e., not a line beginning with
`$`), it produces a `CodeLine`. If there is an empty line,
`remark-code-snippet` produces a `CodeLine` with no text content, which
does not display when the browser renders the `CodeLine`.

To ensure that the `CodeLine` displays properly, this change edits
`remark-code-snippet` to push a `<br>` element to the MDX.js AST it
creates after the empty `CodeLine`. This does not affect copy/paste,
since the copy/paste logic only handles `CodeLine`s.