Fix the "Lint code base" workflow and package.json
This change cleans up the "Lint code base" GitHub Actions workflow in
order to ensure that it works as expected:

- Change `node-version` to 18 to match Vercel as well as the
  `gravitational/teleport` GitHub Actions runner.
- Build the docs to add an extra layer of protection against faulty docs
  changes making it to Vercel
- Remove cacheing logic. `gravitational/docs` PRs aren't frequent enough
  to warrant a cache for node_modules and yarn state, and often need to
  change dependencies anyway. Removing this takes away unneeded
  complexity.  Also worth noting that the cacheing logic referred to an
  `nvm` step that isn't executed in this job.

This change also fixes the "engines" key in `package.json`. This was
previously given as "engine", which is not a valid `package.json` key.
Also constrains the Node version to minor/patch releases of v18.