Remove version warnings in Tabs components (#404)
The Tabs component displays the VersionWarning component instead of the
TabItem configured by the page if a TabItem's scope is set to `cloud`
and the current docs version is not the same as the latest docs version.

This warning is unhelpful, since the "latest" version configured for the
docs site is often one major version ahead of the Teleport Cloud
version. In addition, since we have removed scope switching from the
docs engine, the `scope` prop is supposed to be no-op until we remove
all instances of this prop from the docs site.

Also updates yarn.lock, which was out of date in earlier commits.