Support anchor-only links within partials
Let's say a partial includes a link like this:

```markdown
[Link text](#anchor)
```

The link will render like this, which is incorrect:

```markdown
[Link text](includes/#anchor)
```

This is because of the way `remark-includes` processes relative links in
partials.

Some partials need to link to an anchor within the same page. This
change supports this by ignoring links with references that begin with
`#`. Since all headings are flattened into the containing page, we can
safely ignore file paths for anchor links.