Support Vars in HLJS snippets with no spans
Fixes gravitational/teleport#35907

The `rehype-hljs-var` plugin replaces `Var` tags with placeholder
strings, applies syntax highlighting, then substitutes the placeholder
strings with the original `Var` tags. It assumes that each `Var` must be
the child of a `span` tag, though this assumption is not always correct.
This change removes the requirement that the plugin check children of
`span` tags for `Var`s.