Remove scope icons
We are de-emphasizing Teleport Community Edition and framing Teleport
Enterprise as a single edition, rather than splitting Enterprise and
Enterprise Cloud. As a result, it doesn't make as much sense as it used
to indicate whether a given feature applies to one of three editions.

This change removes the "Available for" icons and navigation icons that
indicate edition availability. This paves the way for information
architecture changes that position Teleport Enterprise as the main
subject of the Teleport docs.