Add a remark plugin for generating a TOC menu (#483)
As we reorganize the documentation, it becomes cumbersome to manually
change lists of links to pages within the documentation, e.g., in table
of contents pages for subsections of the docs. This change adds a remark
plugin for generating a list of links to pages in the current directory.

The plugin works similarly to `remark-includes`, and accesses the local
filesystem during the docs build. It replaces any lines consisting of
`(!toc!)` with a list of links to pages in the current directory. The
assumption is that a category page within a directory can use this to
list contents.