Require Docusaurus-friendly section index pages
If a directory has no section index page within it that is named after
the directory, e.g., `dir1/dir2/dir2.mdx`, throw an exception when
building the docs.

Currently, the docs engine also allows a section index page at the same
level as its corresponding directory, since this is a convention the
docs content has followed. However, this leads to undesired output when
building the site with Docusaurus.