Provisionally update the docs for v17 (#504)
The docs versioning policy for v17 onward is to switch the default docs
version to the newly released major version when we roll it out to Cloud
customers. However, this means that once we release the new major
version, users who visit the tab for that version see a message that the
version is an upcoming release.

This change edits the docs engine to not display the "upcoming release"
message for v17, while keeping v16 as the default docs version. It also
adds an "Edge" submodule for `master`. Since we no longer support v14,
this change removes the v14 submodule.